pyproject.toml
[tool.poetry]
name = "and-platform"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "and_platform"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
Flask-SQLAlchemy = "^3.0.5"
fabric = "^3.1.0"
flask-migrate = "^4.0.4"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
pyjwt = "^2.8.0"
paramiko = "^3.3.1"
flask-jwt-extended = "^4.5.2"
pyyaml = "^6.0.1"
celery = "^5.3.1"
redis = "^5.0.0"
fulgens = {git = "https://github.com/CTF-Compfest-15/fulgens.git", rev = "v0.2.1"}
requests = "^2.31.0"
pytz = "^2023.3"
flask-cors = "^4.0.0"
argparse = "^1.4.0"
gunicorn = "^21.2.0"
gevent = "^23.9.0"
flask-caching = "^2.0.2"
markdown = "^3.4.4"
flask-socketio = "^5.3.6"
gevent-websocket = "^0.10.1"
py-gfm = "^2.0.0"
apscheduler = "^3.10.4"
pika = "^1.3.2"
marshmallow-sqlalchemy = "^1.0.0"
flask-limiter = "^3.8.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.4.1"
flake8 = "^6.1.0"
isort = "^5.12.0"
types-pyyaml = "^6.0.12.11"
pytest = "^8.3.1"
pytest-rabbitmq = "^3.1.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
